[Fix]-Correciones

+++ platzi_market/src/main/java/com/market/platzi_market/web/controllers/ProductController.java
@@ -8,7 +8,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController()
 @RequestMapping("/product")
@@ -20,13 +22,17 @@ public class ProductController {
 
     @GetMapping("")
     private ResponseEntity<List<Product_mysql>> getAll(){
-        return new ResponseEntity<>(productService.getAll(), HttpStatus.OK);
+         List<Product_mysql> datos = productService.getAll();
+        return !datos.isEmpty()
+                ? ResponseEntity.ok(datos)
+                : ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
     }
 
     //por params
     @GetMapping("/{id_categoria}")
     private ResponseEntity <List<Product_mysql>> getByCatgoriaId(@PathVariable("id_categoria") int id_categoria){
-        return new ResponseEntity<>(productService.getProductsByCategory(id_categoria),HttpStatus.OK);
+        List<Product_mysql> datos = productService.getProductsByCategory(id_categoria);
+        return !datos.isEmpty() ? ResponseEntity.ok(datos) : ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
     }
 
     //por query
@@ -38,11 +44,23 @@ public class ProductController {
 
     @PostMapping("")
     private ResponseEntity<Product_mysql> save (@RequestBody Product_mysql product_mysql){
-            return  new ResponseEntity<>(productService.save(product_mysql),HttpStatus.CREATED);
+            return new ResponseEntity<>(productService.save(product_mysql),HttpStatus.CREATED);
+    }
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<Map<String, Object>> delete(@PathVariable Long id) {
+        boolean exist = productService.ExistCategoria(id);
+        if (exist) {
+            productService.delete(id);
+            Map<String, Object> response = new HashMap<>();
+            response.put("ok", true);
+            response.put("msg", "Producto eliminado correctamente");
+            return ResponseEntity.ok(response);
+        } else {
+            Map<String, Object> response = new HashMap<>();
+            response.put("ok", false);
+            response.put("msg", "No existe el producto a eliminar");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+        }
     }
 
-    @DeleteMapping("/delete/{id}")
-    private void  delete(@PathVariable Long  id){
-        productService.delete(id);
-    }
 }
